+++ src/identity/action/mod.rs
@@ -159,6 +159,10 @@ pub enum Action {
         digest: Digest,
         #[serde(skip_serializing_if = "Option::is_none")]
         artifact_type: Option<MediaType>,
+        #[serde(skip_serializing_if = "Option::is_none")]
+        n: Option<u16>,
+        #[serde(skip_serializing_if = "Option::is_none")]
+        last: Option<String>,
     },
     #[serde(rename = "list-revisions")]
     ListRevisions {